Microbial Abundances

Working with microbial abundances

Tables of abundances are based off ComMatrix types from SpatialEcology.jl, where columns are samples and rows are features (eg species). Sample and feature names are also stored, and there's a convenience function if you want to convert a DataFrame to a ComMatrix, assuming the first column contains feature names:

julia> using Microbiome, DataFrames;

julia> df = DataFrame(species=["E. coli", "B. fragilis", "L. casei"],
                             sample1=[1, 4, 5],
                             sample2=[3, 8, 0],
                             sample3=[0, 3, 4]);

julia> abund = abundancetable(df)
ComMatrix with 3 species in 3 sites

Species names:
E. coli, B. fragilis, L. casei

Site names:
sample1, sample2, sample3

Forgive the clutter... ComMatricies name rows as species (which is true in this case, but need not be), and columns are "sites" rather than samples. That will be fixed eventually.

samplenames(abund)
featurenames(abund)
sampletotals(abund) # this is column sums
featuretotals(abund) # this is row sums

3-element Array{Float64,1}:
  4.0
 15.0
  9.0

If you want relative abundance, you can do relativeabundance(abund) or relativeabundance!(abund):

relativeabundance!(abund);

sampletotals(abund)

3-element Array{Float64,1}:
 1.0
 1.0
 1.0

You can also filter on the n most abundant features accross the dataset. This function automatically generates an n+1 row for other containing the remaining features. Note - these doesn't modify in-place, so you've gotta reassign if you want to update:

abund2 = filterabund(abund, 1);

featurenames(abund2)

2-element Array{String,1}:
 "B. fragilis"
 "other"      

Methods

Microbiome.abundancetable — Method.

abundancetable(df::DataFrame)
abundancetable(table::AbstractArray{T,2})

Convert DataFrame or matrix into a ComMatrix

source

Microbiome.colnormalize! — Method.

colnormalize!(abt::AbstractComMatrix)

Normalize rows of a ComMatrix to the maximum of each column.

source

Microbiome.colnormalize — Method.

colnormalize(abt::AbstractComMatrix)

Return a copy of a ComMatrix normalized to the maximum of each column.

source

Microbiome.filterabund — Function.

filterabund(abun::AbstractComMatrix, n::Int=minimum(10, nfeatures(abun)))

Filter an abundance table to the top n features accross all samples

This function also adds a row for "other", which sums the abundances of the remaining features.

source

Microbiome.present — Function.

present(t::Union{Float64, Missing}, minabundance::Float64=0.0001)

Check if a given (non-zero) value is greater than a minimum value. If the minimum abundance is 0, just checks if value is non-zero.

source

Microbiome.prevalence — Function.

prevalence(a::AbstractArray{<:Real}, minabundance::Float64=0.0001)

Return the fraction of values that are greater than a minimum.

source

Microbiome.relativeabundance! — Method.

relativeabundance!(a::AbstractComMatrix; kind::Symbol=:fraction)

Normalize each column of a ComMatrix to the sum of the column.

By default, columns sum to 1.0. Use kind=:percent for columns to sum to 100.

source

Microbiome.relativeabundance — Method.

relativeabundance!(a::AbstractComMatrix; kind::Symbol=:fraction)

Return a copy of a ComMatrix with columns normalized the sum of each column.

By default, columns sum to 1.0. Use kind=:percent for columns to sum to 100.

source

Microbiome.rownormalize! — Method.

rownormalize!(abt::AbstractComMatrix)

Normalize rows of a ComMatrix to the maximum of each row.

source

Microbiome.rownormalize — Method.

rownormalize(abt::AbstractComMatrix)

Return a copy of a ComMatrix normalized to the maximum of each row.

source