Working with Distances / Dissimilarity

Quite often, it's useful to boil stuff down to distances between samples. AbundanceTables can be used with the pairwise() function from Distances.jl to get a symetric distance matrix.

At the moment, this only works if you're on SpatialEcology.jl's master branch.

using Distances
using Microbiome

abund = abundancetable([1  3  0;
                        4  8  3;
                        5  0  4]);

dm = pairwise(BrayCurtis(), abund, dims=2)

Methods

Microbiome.braycurtis — Method

braycurtis(abt::AbstractAbundanceTable)

Retruns a pairwise Bray-Curtis dissimilarity matrix.

Microbiome.ginisimpson! — Method

ginisimpson!(abt::AbstractAbundanceTable; overwrite=false)

Adds a :ginisimpson entry to the metadata for each sample in abt with the Gini-Simpson alpha diversity of that sample (see ginisimpson). If overwrite=false (the default), uses insert! to perform this operation, so an error will be thrown if any sample already contains a :ginisimpson entry. Otherwise, uses set!.

Microbiome.ginisimpson — Method

ginisimpson(v::Union{AbstractVector, AbstractSparseMatrix}) 
ginisimpson(abt::AbstractAbundanceTable, overwrite=false)

Computes the Gini-Simpson alpha diversity metric for a vector. When called on an AbstractAbundanceTable, returns a 1 x nsamples matrix with 1 entry per sample. See also ginisimpson!.

Microbiome.shannon! — Method

shannon!(abt::AbstractAbundanceTable; overwrite=false)

Adds a :shannon entry to the metadata for each sample in abt with the Shannon alpha diversity of that sample (see shannon). If overwrite=false (the default), uses insert! to perform this operation, so an error will be thrown if any sample already contains a :shannon entry. Otherwise, uses set!.

Microbiome.shannon — Method

shannon(v::Union{AbstractVector, AbstractSparseMatrix}) 
shannon(abt::AbstractAbundanceTable)

Computes the Shannon alpha diversity metric for a vector. When called on an AbstractAbundanceTable, returns a 1 x nsamples matrix with 1 entry per sample. See also shannon!.