Working with microbial abundances

The primary type for working with microbial abundances is the CommunityProfile, which is a sparse matrix with MicrobiomeSamples as column indices and features (eg Taxons or GeneFunctions) as row indices.

For example, let's make a CommunityProfile with 3 samples, 5 species and 5 genera.

julia> samps = MicrobiomeSample.(["s1", "s2", "s3"])
3-element Vector{MicrobiomeSample}:
 MicrobiomeSample("s1", {})
 MicrobiomeSample("s2", {})
 MicrobiomeSample("s3", {})

julia> taxa = [[Taxon("s$i", :species) for i in 1:5]; [Taxon("g$i", :genus) for i in 1:5]]
10-element Vector{Taxon}:
 Taxon("s1", :species)
 Taxon("s2", :species)
 Taxon("s3", :species)
 Taxon("s4", :species)
 Taxon("s5", :species)
 Taxon("g1", :genus)
 Taxon("g2", :genus)
 Taxon("g3", :genus)
 Taxon("g4", :genus)
 Taxon("g5", :genus)

julia> mat = spzeros(10, 3); # 10 x 3 matrix filled with zeros

julia> for i in 1:10, j in 1:3 
           # fill some spots with random values
           rand() < 0.3 && (mat[i,j] = rand())
       end

julia> mat
10×3 SparseMatrixCSC{Float64, Int64} with 10 stored entries:
  ⋅         ⋅        0.172933
  ⋅         ⋅         ⋅
 0.956916   ⋅         ⋅
 0.422956   ⋅         ⋅
  ⋅         ⋅         ⋅
 0.502952   ⋅        0.167169
  ⋅         ⋅        0.244683
  ⋅         ⋅        0.143638
 0.570085  0.249238   ⋅
  ⋅        0.841643   ⋅

julia> cp = CommunityProfile(mat, taxa, samps)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 features in 3 samples

Feature names:
s1, s2, s3...g4, g5

Sample names:
s1, s2, s3

Accessing CommunitProfile contents

It is easy to get out the underlying abundance matrix, features, and samples using abundances, features, and samples respectively:

julia> abundances(cp)
10×3 SparseMatrixCSC{Float64, Int64} with 10 stored entries:
  ⋅         ⋅        0.172933
  ⋅         ⋅         ⋅
 0.956916   ⋅         ⋅
 0.422956   ⋅         ⋅
  ⋅         ⋅         ⋅
 0.502952   ⋅        0.167169
  ⋅         ⋅        0.244683
  ⋅         ⋅        0.143638
 0.570085  0.249238   ⋅
  ⋅        0.841643   ⋅

julia> features(cp)
10-element Vector{Taxon}:
 Taxon("s1", :species)
 Taxon("s2", :species)
 Taxon("s3", :species)
 Taxon("s4", :species)
 Taxon("s5", :species)
 Taxon("g1", :genus)
 Taxon("g2", :genus)
 Taxon("g3", :genus)
 Taxon("g4", :genus)
 Taxon("g5", :genus)
 
julia> samples(cp)
3-element Vector{MicrobiomeSample}:
 MicrobiomeSample("s1", {})
 MicrobiomeSample("s2", {})
 MicrobiomeSample("s3", {})

You can also just get an array of feature and sample names using featurenames and samplenames respectively.

julia> featurenames(cp)
10-element Vector{String}:
 "s1"
 "s2"
 "s3"
 "s4"
 "s5"
 "g1"
 "g2"
 "g3"
 "g4"
 "g5"

julia> samplenames(cp)
3-element Vector{String}:
 "s1"
 "s2"
 "s3"

Finally, you can pull out the metadata of all samples or a subset using metadata. The returned value is a vector of NamedTuples, which is compliant with the Tables.jl interface, so it's easy to load into other formats (like DataFrames.jl for example):

julia> metadata(cp)
3-element Vector{NamedTuple{(:sample,), Tuple{String}}}:
 (sample = "s1",)
 (sample = "s2",)
 (sample = "s3",)

Of course, for now, the metadata is pretty boring - jump ahead to Working with metadata to do some more fun things.

Indexing and selecting

CommunityProfiles wrap a sparse matrix, and you can access the values as you would a normal matrix. In julia, you can pull out specific values using [row, col]. So for example, to get the 3rd row, 2nd column, of matrix mat:

julia> mat = reshape(1:12, 4, 3) |> collect
4×3 Matrix{Int64}:
 1  5   9
 2  6  10
 3  7  11
 4  8  12

julia> mat[3,2]
7

You can also get "slices", eg to get rows 2-4, column 1:

julia> mat[2:4, 1]
3-element Vector{Int64}:
 2
 3
 4

To get all of one dimension, you can just use a bare :

julia> mat[:, 1:2]
4×2 Matrix{Int64}:
 1  5
 2  6
 3  7
 4  8

For CommunityProfiles, indexing with integer values will return the value of the matrix at that position, while indexing with slices will return a new CommunityProfile:

julia> cp[1,3]
0.17293302893695128

julia> cp[1:3,6]
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 3 features in 1 samples

Feature names:
s1, s2, s3

Sample names:
g1

julia> cp[1:3,3] |> abundances
3×1 SparseMatrixCSC{Float64, Int64} with 1 stored entry:
 0.172933
  ⋅
  ⋅

Indexing with strings and regular expressions

It is often inconvenient to find the numerical index of a particular feature or sample. Instead, you can use strings or regular expressions to get slices of a CommunityProfile, which will match on the name field of the features or samples. This kind of indexing always returns a CommunityProfile, even if it only has 1 value.

julia> cp["g1", "s1"]
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 1 features in 1 samples

Feature names:
g1

Sample names:
s1

julia> cp[r"[gs]1", "s1"]
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 2 features in 1 samples

Feature names:
s1, g1

Sample names:
s1

Working with metadata

The metadata dictionaries of MicrobiomeSamples can accessed and updated inside a CommunityProfile using insert!, delete!, set!, and unset!. either by directly acting on the underlying sample object, or using special methods that take the sample name as the second argument:

julia> set!(samples(cp)[1], :subject, "kevin")
MicrobiomeSample("s1", {:subject = "kevin"})

julia> insert!(cp, "s2", :subject, "anika")
MicrobiomeSample("s2", {:subject = "anika"})

You can also retrieve all metadata associated with a table using metadata, which will return a Tables.jl-compatible vector or NamedTuples, where each "row" corresponds to one sample. All metadata fields found in any sample will be returned in every row, with the value missing in any samples that do not have that field set.

julia> metadata(cp)
3-element Vector{NamedTuple{(:sample, :subject), T} where T<:Tuple}:
 (sample = "s1", subject = "kevin")
 (sample = "s2", subject = "anika")
 (sample = "s3", subject = missing)

And you can bulk-insert! or set! metadata by passing a similar Table-like object with the a field (:sample by default) matching sample names found in the CommunityProfile.

julia> md = [(sample="s1", subject="kevin", foo="bar"), (sample="s3", subject="annelle", foo="baz")]
2-element Vector{NamedTuple{(:sample, :subject, :foo), Tuple{String, String, String}}}:
 (sample = "s1", subject = "kevin", foo = "bar")
 (sample = "s3", subject = "annelle", foo = "baz")

julia> set!(cp, md)

julia> md2 = [(name="s1", other="Hello, World!"), (name="s2", other="Goodbye!")]
2-element Vector{NamedTuple{(:name, :other), Tuple{String, String}}}:
 (name = "s1", other = "Hello, World!")
 (name = "s2", other = "Goodbye!")

julia> insert!(cp, md2; namecol=:name)

julia> metadata(cp)
3-element Vector{NamedTuple{(:sample, :subject, :foo, :other), T} where T<:Tuple}:
 (sample = "s1", subject = "kevin", foo = "bar", other = "Hello, World!")
 (sample = "s2", subject = "anika", foo = missing, other = "Goodbye!")
 (sample = "s3", subject = "annelle", foo = "baz", other = missing)

Types and Methods

Microbiome.CommunityProfile — Type

CommunityProfile{T, F, S} <: AbstractAbundanceTable{T, F, S}

An AbstractAssemblage from EcoBase.jl that uses an AxisArray of a SparseMatrixCSC under the hood.

CommunityProfiles are tables with AbstractFeature-indexed rows and AbstractSample-indexed columns. Note - we can use the name of samples and features to index. ```

Microbiome.samples — Function

samples(at::AbstractAbundanceTable)

Returns samples in at. To get samplenames instead, use samplenames.

samples(at::AbstractAbundanceTable, name::AbstractString)

Returns sample in at with name name.

Microbiome.features — Function

features(at::AbstractAbundanceTable)

Returns features in at. To get featurenames instead, use featurenames.

Microbiome.samplenames — Function

samplenames(at::AbstractAbundanceTable)

Get a vector of sample names from at, equivalent to name.(samples(at))

Microbiome.featurenames — Function

featurenames(at::AbstractAbundanceTable)

Get a vector of feature names from at, equivalent to name.(features(at))

Microbiome.commjoin — Function

commjoin(c1::CommunityProfile, comms::CommunityProfile...)

Join multiple CommunityProfiles, creating a new CommunityProfile. For now, sample names cannot overlap in any of the input profiles.

Microbiome.relativeabundance — Function

relativeabundance(at::AbstractAbundanceTable, kind::Symbol=:fraction)

Like relativeabundance!, but does not mutate original.

Microbiome.relativeabundance! — Function

relativeabundance!(a::AbstractAbundanceTable; kind::Symbol=:fraction)

Normalize each sample in AbstractAbundanceTable to the sum of the sample.

By default, columns sum to 1.0. Use kind=:percent for columns to sum to 100.

Microbiome.present — Function

present(t::Union{Real, Missing}, minabundance::Real=0.0)
present(at::AbstractAbundanceTable, minabundance::Real=0.0)

Check if a given (non-zero) value is greater than or equal to a minimum value. If the minimum abundance is 0, just checks if value is non-zero.

If used on an AbstractAbundanceTable, returns a sparse boolean matrix of the same size.

Microbiome.prevalence — Function

prevalence(a::AbstractArray{<:Real}, minabundance::Real=0.0)
prevalence(at::AbstractAbundanceTable, minabundance::Real=0.0)

Return the fraction of values that are greater than or equal to a minimum. If the minimum abundance is 0, returns the fraction of non-zero values.

If used on an AbstractAbundanceTable, returns a prevalence value for each feature accross the samples.

Microbiome.prevalence_filter — Function

prevalence_filter(comm::AbstractAbundanceTable; minabundance=0.0; minprevalence=0.05, renorm=false)

Return a filtered CommunityProfile where features with prevalence lower than minprevalence are removed. By default, a feature is considered "present" if > 0, but this can be changed by setting minabundance.

Optionally, set renorm = true to calculate relative abundances after low prevalence features are removed.