Working with microbial abundances

Tables of abundances are based off ComMatrix types from SpatialEcology.jl, where columns are samples and rows are features (eg species). Sample and feature names are also stored, and there's a convenience function if you want to convert a DataFrame to a ComMatrix, assuming the first column contains feature names:

julia> using Microbiome, DataFrames;
ERROR: ArgumentError: Package DataFrames not found in current path:
- Run `import Pkg; Pkg.add("DataFrames")` to install the DataFrames package.

julia> df = DataFrame(species=["E. coli", "B. fragilis", "L. casei"],
                             sample1=[1, 4, 5],
                             sample2=[3, 8, 0],
                             sample3=[0, 3, 4]);
ERROR: UndefVarError: DataFrame not defined

julia> abund = abundancetable(df)
ERROR: UndefVarError: abundancetable not defined

Forgive the clutter... ComMatricies name rows as species (which is true in this case, but need not be), and columns are "sites" rather than samples. That will be fixed eventually.

julia> samplenames(abund)
ERROR: UndefVarError: abund not defined

julia> featurenames(abund)
ERROR: UndefVarError: abund not defined

julia> sampletotals(abund) # this is column sums
ERROR: UndefVarError: abund not defined

julia> featuretotals(abund) # this is row sums
ERROR: UndefVarError: abund not defined

If you want relative abundance, you can do relativeabundance(abund) or relativeabundance!(abund):

julia> relativeabundance!(abund);
ERROR: UndefVarError: abund not defined

julia> sampletotals(abund)
ERROR: UndefVarError: abund not defined

You can also filter on the n most abundant features accross the dataset. This function automatically generates an n+1 row for other containing the remaining features. Note - these doesn't modify in-place, so you've gotta reassign if you want to update:

julia> abund2 = filterabund(abund, 1);
ERROR: UndefVarError: filterabund not defined

julia> featurenames(abund2)
ERROR: UndefVarError: abund2 not defined

Methods

Microbiome.present — Function

present(t::Union{Real, Missing}, minabundance::Real=0.0)
present(at::AbstractAbundanceTable, minabundance::Real=0.0)

Check if a given (non-zero) value is greater than or equal to a minimum value. If the minimum abundance is 0, just checks if value is non-zero.

If used on an AbstractAbundanceTable, returns a sparse boolean matrix of the same size.

Microbiome.prevalence — Function

prevalence(a::AbstractArray{<:Real}, minabundance::Real=0.0)
prevalence(at::AbstractAbundanceTable, minabundance::Real=0.0)

Return the fraction of values that are greater than or equal to a minimum. If the minimum abundance is 0, returns the fraction of non-zero values.

If used on an AbstractAbundanceTable, returns a prevalence value for each feature accross the samples.

Microbiome.relativeabundance! — Method

relativeabundance!(a::AbstractAbundanceTable; kind::Symbol=:fraction)

Normalize each sample in AbstractAbundanceTable to the sum of the sample.

By default, columns sum to 1.0. Use kind=:percent for columns to sum to 100.

Microbiome.relativeabundance — Method

relativeabundance(at::AbstractAbundanceTable, kind::Symbol=:fraction)

Like relativeabundance!, but does not mutate original.