Working with microbial abundances

Microbiome.CommunityProfile — Type

CommunityProfile{T, F, S} <: AbstractAbundanceTable{T, F, S}

An AbstractAssemblage from EcoBase.jl that uses an AxisArray of a SparseMatrixCSC under the hood.

CommunityProfiles are tables with AbstractFeature-indexed rows and AbstractSample-indexed columns. Note - we can use the name of samples and features to index.

julia> txs = [Taxon("taxon$i") for i in 1:10];

julia> mss = [MicrobiomeSample("sample$i") for i in 1:5];

julia> mat = spzeros(10,5);

julia> for i in 1:5; mat[i,i] = 1.; end

julia> comm = CommunityProfile(mat, txs, mss)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 features in 5 samples

Feature names:
taxon1, taxon2, taxon3...taxon9, taxon10

Sample names:
sample1, sample2, sample3, sample4, sample5

julia> comm["taxon1", "sample1"]
1.0

julia> comm[:,["sample1", "sample5"]]
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 features in 2 samples

Feature names:
taxon1, taxon2, taxon3...taxon9, taxon10

Sample names:
sample1, sample5

julia> comm[Taxon("taxon3", :kingdom), "sample1"]
0.0

Microbiome.samples — Function

samples(at::AbstractAbundanceTable)

Returns samples in at. To get samplenames instead, use samplenames.

samples(at::AbstractAbundanceTable, name::AbstractString)

Returns sample in at with name name.

Microbiome.features — Function

features(at::AbstractAbundanceTable)

Returns features in at. To get featurenames instead, use featurenames.

Microbiome.samplenames — Function

samplenames(at::AbstractAbundanceTable)

Get a vector of sample names from at, equivalent to name.(samples(at))

Microbiome.featurenames — Function

featurenames(at::AbstractAbundanceTable)

Get a vector of feature names from at, equivalent to name.(features(at))

Microbiome.commjoin — Function

commjoin(c1::CommunityProfile, comms::CommunityProfile...)

Join multiple CommunityProfiles, creating a new CommunityProfile. For now, sample names cannot overlap in any of the input profiles.

julia> mss = [MicrobiomeSample(string("sample",i)) for i in 1:15];

julia> txs = [Taxon(string("taxon",i)) for i in 1:20];

julia> cm1 = CommunityProfile(spzeros(10,5), txs[1:10], mss[1:5])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 features in 5 samples

Feature names:
taxon1, taxon2, taxon3...taxon9, taxon10

Sample names:
sample1, sample2, sample3, sample4, sample5

julia> cm2 = CommunityProfile(spzeros(10,5), txs[6:15], mss[6:10])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 features in 5 samples

Feature names:
taxon6, taxon7, taxon8...taxon14, taxon15

Sample names:
sample6, sample7, sample8, sample9, sample10

julia> cm3 = CommunityProfile(spzeros(10,5), txs[11:20], mss[11:15])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 features in 5 samples

Feature names:
taxon11, taxon12, taxon13...taxon19, taxon20

Sample names:
sample11, sample12, sample13, sample14, sample15

julia> commjoin(cm1, cm2, cm3)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 20 features in 15 samples

Feature names:
taxon1, taxon2, taxon3...taxon19, taxon20

Sample names:
sample1, sample2, sample3...sample14, sample15

Microbiome.relativeabundance — Function

relativeabundance(at::AbstractAbundanceTable, kind::Symbol=:fraction)

Like relativeabundance!, but does not mutate original.

Microbiome.relativeabundance! — Function

relativeabundance!(a::AbstractAbundanceTable; kind::Symbol=:fraction)

Normalize each sample in AbstractAbundanceTable to the sum of the sample.

By default, columns sum to 1.0. Use kind=:percent for columns to sum to 100.

Microbiome.present — Function

present(t::Union{Real, Missing}, minabundance::Real=0.0)
present(at::AbstractAbundanceTable, minabundance::Real=0.0)

Check if a given (non-zero) value is greater than or equal to a minimum value. If the minimum abundance is 0, just checks if value is non-zero.

If used on an AbstractAbundanceTable, returns a sparse boolean matrix of the same size.

Microbiome.prevalence — Function

prevalence(a::AbstractArray{<:Real}, minabundance::Real=0.0)
prevalence(at::AbstractAbundanceTable, minabundance::Real=0.0)

Return the fraction of values that are greater than or equal to a minimum. If the minimum abundance is 0, returns the fraction of non-zero values.

If used on an AbstractAbundanceTable, returns a prevalence value for each feature accross the samples.

Microbiome.prevalence_filter — Function

prevalence_filter(comm::AbstractAbundanceTable; minabundance=0.0; minprevalence=0.05, renorm=false)

Return a filtered CommunityProfile where features with prevalence lower than minprevalence are removed. By default, a feature is considered "present" if > 0, but this can be changed by setting minabundance.

Optionally, set renorm = true to calculate relative abundances after low prevalence features are removed.

julia> comm = CommunityProfile(sparse([3 0 1 # 0.33, assuming minabundance 2
                                       2 2 2 # 1.0
                                       0 0 1 # 0.0
                                       2 0 0 # 0.33
                                       ]),
                               [Taxon(string(i)) for i in 1:4],
                               [MicrobiomeSample(string(i)) for i in 1:3]);

julia> prevalence_filter(comm, minabundance=2, minprevalence=0.3) 
CommunityProfile{Int64, Taxon, MicrobiomeSample} with 3 features in 3 samples

Feature names:
1, 2, 4

Sample names:
1, 2, 3

julia> prevalence_filter(comm, minabundance=2, minprevalence=0.4)
CommunityProfile{Int64, Taxon, MicrobiomeSample} with 1 features in 3 samples

Feature names:
2

Sample names:
1, 2, 3

julia> prevalence_filter(comm, minabundance=3, minprevalence=0.3)
CommunityProfile{Int64, Taxon, MicrobiomeSample} with 1 features in 3 samples

Feature names:
1

Sample names:
1, 2, 3