Samples and Features

Microbial profiles are made up of AbstractSamples and AbstractFeatures. Typically, an AbstractSample is an individual biospecimen or other observation, and contains some number of AbstractFeatures, such as taxa or gene functions. AbstractSamples may also contain arbitrary metadata.

Sample types

At its most basic, an AbstractSample simply encodes a name (which should be a unique identifier), and a place to hold metadata. The concrete type MicrobiomeSample is implemented with these two fields, the latter of which is a Dictionary from Dictionaries.jl.

Microbiome.MicrobiomeSample — Type

MicrobiomeSample(name::String, metadata::Dictionary{Symbol, T}) <: AbstractSample
MicrobiomeSample(name::String)

Microbiome sample type that includes a name and a Dictionary of arbitrary metadata using Symbols (other than :name or :metadata) as keys.

Metadata can be accessed using getproperty or getindex on the sample itself.

julia> ms = MicrobiomeSample("sample1", Dictionary([:gender, :age], ["female", 180]))
MicrobiomeSample("sample1", {:gender │ "female", :age │ 180})

julia> name(ms)
"sample1"

julia> ms.name
"sample1"

julia> ms.gender
"female"

julia> ms.age
180

Samples can be instantiated with only a name, leaving the metadata Dictionary blank

julia> ms2 = MicrobiomeSample("sample2")
MicrobiomeSample("sample2", {})

Adding or changing metadata follows the same rules as for the normal Dictionary type.

to change a value use setproperty or setindex. Note that this will fail if the key does not already exist.

to add or remove a value with validation that it does or does not exist already, use insert! and delete! respectively.

to add or remove a value without validation ("upsert"), use set! and unset! respectively.

julia> ms
MicrobiomeSample("sample1", {:gender │ "female", :age │ 180})

julia> ms.age = 16 * 365 # or `ms[:age] = 16 * 365`
5840

julia> set!(ms, :gender, "nonbinary")
MicrobiomeSample("sample1", {:gender │ "nonbinary", :age │ 5840})

julia> insert!(ms, :occupation, "clerk")
MicrobiomeSample("sample1", {:gender │ "nonbinary", :age │ 5840, :occupation │ "clerk"})

julia> insert!(ms, :occupation, "bagger")
ERROR: IndexError("Dictionary already contains index: occupation")

julia> delete!(ms, :occupation)
MicrobiomeSample("sample1", {:gender │ "nonbinary", :age │ 5840})

julia> delete!(ms, :occupation)
ERROR: IndexError("Index doesn't exist: occupation")

julia> unset!(ms, :occupation)
MicrobiomeSample("sample1", {:gender │ "nonbinary", :age │ 5840})

Feature Types

AbstractFeature types also have a name, but other fields are optional. Microbiome.jl defines two concrete AbstractFeature types, Taxon and GeneFunction.

Microbiome.Taxon — Type

Taxon(name::String, clade::Union{Missing, Symbol, Int}) <: AbstractFeature
Taxon(name::String)

Microbial taxon with a name and a clade that can be one of

:domain

:kingom

:phylum

:class

:order

:faamily

:genus

:species

:subspecies

:strain

or missing. Contructors can also use numbers 0-9, or pass a string alone (in which case the taxon will be stored as missing)

Microbiome.GeneFunction — Type

GeneFunction(name::String, taxon::Union{Taxon, String, Missing}) <: AbstractFeature
GeneFunction(name::String)

Microbial gene function object with optional stratification (taxon).

Types and Methods

Microbiome.metadata — Function

metadata(t::AbstractSample)

Get the metadata field from an AbstractSample. Note that this is not a copy, so modifications to the returned value will update the parent AbstractSample as well.

Microbiome.name — Function

name(t::Union{AbstractSample, AbstractFeature})

Get the name field from an AbstractSample or AbstractFeature.

Microbiome.clade — Function

clade(t::Union{Taxon, missing})

Get the clade field from an Taxon. Returns missing if the clade is not set.

Microbiome.hasclade — Function

hasclade(t::Taxon)::Bool

Pretty self-explanatory.

Microbiome.taxon — Function

taxon(t::Union{GeneFunction, missing})

Get the taxon field from a GeneFunction. Returns missing if the taxon is not set.

Microbiome.hastaxon — Function

hastaxon(t::GeneFunction)::Bool

Pretty self-explanatory.